move Navigation Bar into the Dashboard Layour, make common folder for Dashboard
